unittests, api

+++ simulator/src/main/java/at/jku/API.java
@@ -12,6 +12,7 @@ public interface API {
     public HttpResponse PUT(String uri);
 
     public HttpResponse DELETE(String uri);
+
     public HttpResponse getRooms();
 
     public HttpResponse postRoom();
@@ -20,51 +21,51 @@ public interface API {
 
     public HttpResponse postRoom(String name, int size);
 
-    public HttpResponse getRoom(int roomID);
+    public HttpResponse getRoom(Long roomID);
 
-    public HttpResponse putRoom(int roomID, String name, int size);
+    public HttpResponse putRoom(Long roomID, String name, int size);
 
-    public HttpResponse deleteRoom(int roomID);
+    public HttpResponse deleteRoom(Long roomID);
 
-    public HttpResponse getLightSources(int roomID);
+    public HttpResponse getLightSources(Long roomID);
 
-    public HttpResponse postLightSource(int roomID);
+    public HttpResponse postLightSource(Long roomID);
 
-    public HttpResponse getLightSource(int roomID, int lightID);
+    public HttpResponse getLightSource(Long roomID, int lightID);
 
-    public HttpResponse putLightSource(int roomID, int lightID, boolean state);
+    public HttpResponse putLightSource(Long roomID, int lightID, boolean state);
 
-    public HttpResponse getWindows(int roomID);
+    public HttpResponse getWindows(Long roomID);
 
-    public HttpResponse postWindow(int roomID);
+    public HttpResponse postWindow(Long roomID);
 
-    public HttpResponse getWindow(int roomID, int windowID);
+    public HttpResponse getWindow(Long roomID, int windowID);
 
-    public HttpResponse putWindow(int roomID, int windowID, boolean state);
+    public HttpResponse putWindow(Long roomID, int windowID, boolean state);
 
-    public HttpResponse getVentilators(int roomID);
+    public HttpResponse getVentilators(Long roomID);
 
-    public HttpResponse postVentilator(int roomID);
+    public HttpResponse postVentilator(Long roomID);
 
-    public HttpResponse getVentilator(int roomID, int ventilatorID);
+    public HttpResponse getVentilator(Long roomID, int ventilatorID);
 
-    public HttpResponse putVentilator(int roomID, int ventilatorID, boolean state);
+    public HttpResponse putVentilator(Long roomID, int ventilatorID, boolean state);
 
-    public HttpResponse getPeopleInRoom(int roomID);
+    public HttpResponse getPeopleInRoom(Long roomID);
 
-    public HttpResponse postPeopleInRoom(int roomID, LocalDateTime timestamp, int peopleInRoom, boolean state);
+    public HttpResponse postPeopleInRoom(Long roomID, LocalDateTime timestamp, int peopleInRoom, boolean state);
 
-    public HttpResponse getTemperature(int roomID);
+    public HttpResponse getTemperature(Long roomID);
 
-    public HttpResponse postTemperature(int roomID, LocalDateTime timestamp, double temperature, boolean state);
+    public HttpResponse postTemperature(Long roomID, LocalDateTime timestamp, double temperature, boolean state);
 
-    public HttpResponse getCo2(int roomID);
+    public HttpResponse getCo2(Long roomID);
 
-    public HttpResponse postCo2(int roomID, LocalDateTime timestamp, double co2, boolean state);
+    public HttpResponse postCo2(Long roomID, LocalDateTime timestamp, double co2, boolean state);
 
-    public HttpResponse getHumidity(int roomID);
+    public HttpResponse getHumidity(Long roomID);
 
-    public HttpResponse postHumidity(int roomID, LocalDateTime timestamp, double humidity, boolean state);
+    public HttpResponse postHumidity(Long roomID, LocalDateTime timestamp, double humidity, boolean state);
 
 
     // TODO missing ventilator activation (needed?) - can be done with standard put
